Criação das Reviews an Stars Rating ok.

+++ database/migrations/2020_03_05_203929_create_reviews_table.php
@@ -17,7 +17,7 @@ class CreateReviewsTable extends Migration
             $table->uuid('id')->primary();
             $table->unsignedBigInteger('bookable_id')->index();
             $table->unsignedBigInteger('booking_id')->index()->nullable();
-            $table->unsignedTinyInteger('rating');
+            $table->unsignedSmallInteger('rating');
             $table->text('content');
             $table->timestamps();
